Extract ProjectParser
